Refactor to source pattern
