+++ wsidicomizer/bioformats/bioformats_cli.py
@@ -1,17 +1,32 @@
+#    Copyright 2023 SECTRA AB
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+
 from typing import Optional, Sequence
 
 from pydicom import Dataset
-from wsidicomizer.bioformats import BioFormatsDicomizer
-from wsidicomizer.cli.wsidicomizer_cli import WsiDicomizerCli
+from wsidicomizer.bioformats.bioformats_dicomizer import BioformatsDicomizer
+from wsidicomizer.cli import WsiDicomizerCli
 
 
-class BioFormatsCli(WsiDicomizerCli):
+class BioformatsCli(WsiDicomizerCli):
     def convert(
         self,
         filepath: str,
         output_path: str,
         modules: Optional[Dataset] = None,
-        tile_size: Optional[int] = 512,
+        tile_size: int = 512,
         include_levels: Optional[Sequence[int]] = None,
         include_label: bool = True,
         include_overview: bool = True,
@@ -23,7 +38,7 @@ class BioFormatsCli(WsiDicomizerCli):
         jpeg_subsampling: str = '420',
         offset_table: Optional[str] = 'bot'
     ):
-        with BioFormatsDicomizer.open(
+        with BioformatsDicomizer.open(
             filepath=filepath,
             modules=modules,
             tile_size=tile_size,
@@ -43,7 +58,7 @@ class BioFormatsCli(WsiDicomizerCli):
 
 
 def main():
-    cli = BioFormatsCli()
+    cli = BioformatsCli()
     cli.cli()
